pyproject.toml
[project]
name = "image-classifier-model"
version = "0.1.0"
description = "Mobile screenshot classification for Failure/Loading/Success states"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    # Core ML frameworks
    "torch>=2.2.0",
    "torchvision>=0.17.0",
    # Data processing
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    # Model export
    "onnx>=1.15.0",
    "onnxruntime>=1.16.0",
    "coremltools>=7.1",
    # Training utilities
    "tensorboard>=2.15.0",
    "tqdm>=4.66.0",
    "pyyaml>=6.0",
    # Metrics and visualization
    "scikit-learn>=1.3.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    # Code quality
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "onnxscript>=0.5.7",
]

[project.optional-dependencies]
dev = [
    "jupyter>=1.0.0",
    "ipywidgets>=8.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]


line-length = 100
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "notebooks/*"]
